Added unit tests for --create-folders option.
