+++ Rakefile
@@ -2,7 +2,8 @@ require 'rake'
 require 'rake/testtask'
 
 Rake::TestTask.new do |t|
-  t.test_files = Dir.glob 'test/**/*_test.rb'
+  t.libs = ['lib', 'test']
+  t.pattern = 'test/**/*_test.rb'
 end
 
 task :default => :test
