+++ test/fixtures/twine_key1_en_es.txt
@@ -1,0 +1,4 @@
+[[My Strings]]
+  [key1]
+    en = key1-english
+    es = key1-spanish
